Initial Commit
Basic functionality is working, user can log tasks, can use camera to capture image for the task, select date, etc. location is not yet resolved.

User settings configuration not yet implemented.

About to simplify a few things, so this is a commit so that I can restore the code I'm stripping out at a later stage.