Commit for Assessment
My Activities App is now fully functional. Could use a fair bit more polish for an actual release, but it does everything requested of it at this point.
It is optimised for phone sized devices, partly as I felt these were the most likely type to use such an app, but largely as I far more frequently use a phone myself.